# api/openapi-rest.yml
openapi: 3.0.4
info:
  title: Swagger Redis Robin
  description: |-
    Api developed for Redis Robin, a faithful partner who assists the operator in the dangerous Gotham.
  contact:
    email: lddpaas@inditex.com
  version: 1.0.0
...
servers:
  - url: https://api.inditex.com/redisrobin
    description: Production server
...
paths:
  /v1/redkeycluster/status:
    get:
      tags:
        - status
      summary: Get current RedKey Cluster status
      description: Get current RedKey Cluster status from Redis Operator perspective
      operationId: GetRedKeyClusterStatus
      responses:
        '200':
          description: Return current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedKeyClusterStatusResponse'
    put:
      tags:
        - status
      summary: Update RedKey Cluster status
      description: Update RedKey Cluster status
      operationId: UpdateRedKeyClusterStatus
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedKeyClusterStatusRequest'
      responses:
        '200':
          description: Return updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedKeyClusterStatusResponse'
...
  /v1/cluster/move:
    put:
      tags:
        - move
      summary: Move slots between Redis Nodes
      description: Move slots between Redis Nodes
      operationId: MoveNodeSlots
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterMoveSlotsRequest'
      responses:
        '200':
          description: From node is already empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterMoveSlotsResponse'
        '201':
          description: Move operation has been launched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterMoveSlotsResponse'
        '202':
          description: Move operation is already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterMoveSlotsResponse'
...
  /v1/cluster/reset/{nodeIndex}:
    put:
      tags:
        - reset
      summary: Reset Cluster node
      description: Launches a cluster node reset
      operationId: ResetNode
      parameters:
        - in: path
          name: nodeIndex
          schema:
            type: integer
          description: Node ID
          required: true
          example: 0
      responses:
        '200':
          description: Cluster node has been reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterFixResponse'
...
components:
  schemas:
    RedKeyClusterStatusResponse:
      type: object
      properties:
        status:
          type: string
          description: RedKey Cluster Status
          example: Ready
          enum:
            - Initializing
            - Ready
            - Error
            - Upgrading
            - ScalingDown
            - ScalingUp
            - Maintenance
            - Unknown
    RedKeyClusterStatusRequest:
      type: object
      properties:
        status:
          type: string
          description: RedKey Cluster Status
          example: Ready
          enum:
            - Initializing
            - Ready
            - Error
            - Upgrading
            - ScalingDown
            - ScalingUp
            - Maintenance
            - Unknown
...
    ClusterMoveSlotsRequest:
      type: object
      required:
        - from
        - to
      properties:
        from:
          type: string
          description: RedKey Cluster node to move slots from. Can be the name name or the node index
          example: "0"
        to:
          type: string
          description: RedKey Cluster node to move slots to. Can be the name name or the node index
          example: redkeycluster-sample-1
        slots:
          type: integer
          description: Slots to move. If not provided, all slots will be moved
          example: 0
          default: 0
    ClusterMoveSlotsResponse:
      type: object
      properties:
        status:
          type: string
          description: Move operation status
          example: In progress
          enum:
            - In progress
            - Completed
            - Error
...
    ClusterFixResponse:
      type: object
      properties:
        status:
          type: string
          description: RedKey Cluster fix status
          example: InProgress
          enum:
            - InProgress
            - Error
...